Add new 'auto_format_git_diff' behavior
When you edit a file tracked in git, it may happend that some tracked
lines are not well formatted. If you modify this file, and you
auto_format the file, lines with bad indentation will be formatted too.
You will be end up with unrelated changes in your commit.

It may even appear that these 'bad formatted lines' were indented like
this with a purpose.

To mitigate this issue, this commit proposes to only format diff'ed
lines in a file. It roughly relies on git diff, and on clang-format
'-lines' option.

clang_format#format function has been modified to accept a list of
ranges.

It adds three new user options:
- g:clang_format#auto_format_git_diff to enable this new behavior
- g:clang_format#auto_format_git_diff_fallback to decide what to do in
  case of untracked file
- g:clang_format#git to define non standard git path or options