Add clang-format re mapping for =
When g:clang_format#equal_indent is 1, and when clang-format is enabled,
"=" command will use clang-format.
It works for:
={motion} (like =% or =})
==
visual select + =

It must be noted that this is a remap. The clean way would have been to
use equalprg, but it hard to use in this case as equalprg is called for
each line, formatting the current line depending of the previous one.